attempted refactor
